fixes follwing PR comments
