+++ web/client/product/plugins.js
@@ -86,9 +86,8 @@ module.exports = {
         RulesDataGridPlugin: require('../plugins/RulesDataGrid'),
         RulesEditorPlugin: require('../plugins/RulesEditor'),
         RulesManagerFooter: require('../plugins/RulesManagerFooter'),
-        // SaveAsPlugin: require('../plugins/SaveAs'),
-        MapSavePlugin: require('../plugins/MapSave').MapSave,
-        MapSaveAsPlugin: require('../plugins/MapSave').MapSaveAs,
+        SavePlugin: require('../plugins/Save').default,
+        SaveAsPlugin: require('../plugins/SaveAs').default,
         SaveStoryPlugin: require('../plugins/GeoStorySave'),
         ScaleBoxPlugin: require('../plugins/ScaleBox'),
         ScrollTopPlugin: require('../plugins/ScrollTop'),
